+++ src/score/score.js
